hubspot-cron-job/requirements.txt
hubspot-api-client
requests
sqlalchemy
pymysql